requirements.txt
python-dotenv==1.0.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
langchain==0.1.13
langchain-community==0.0.29
langchain-google-genai==1.0.4
pypdf==4.1.0
tiktoken==0.6.0

